msvc/vpxpes_parser: Silence integer conversion warning.

+++ m2ts/vpxpes_parser.cc
@@ -38,7 +38,7 @@ bool VpxPesParser::Open(const std::string& pes_file) {
   pes_file_size_ = libwebm::GetFileSize(pes_file);
   if (pes_file_size_ <= 0)
     return false;
-  pes_file_data_.reserve(pes_file_size_);
+  pes_file_data_.reserve(static_cast<size_t>(pes_file_size_));
   libwebm::FilePtr file = libwebm::FilePtr(std::fopen(pes_file.c_str(), "rb"),
                                            libwebm::FILEDeleter());
